+++ src/rulesets/oas3/__tests__/valid-oas-example-in-content.ts
@@ -1,0 +1,5 @@
+import testRule from './templates/_oas-example';
+
+const ruleName = 'valid-oas-example-in-content';
+
+describe(ruleName, () => testRule(ruleName, 'content'));
