+++ src/rulesets/oas3/__tests__/valid-oas-example-in-headers.ts
@@ -1,276 +1,5 @@
-import { DiagnosticSeverity } from '@stoplight/types';
-import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import testRule from './templates/_oas-example';
 
-// @oclif/test packages requires @types/mocha, therefore we have 2 packages coming up with similar typings
-// TS is confused and prefers the mocha ones, so we need to instrument it to pick up the Jest ones
-declare var test: jest.It;
+const ruleName = 'valid-oas-example-in-headers';
 
-describe('valid-oas-example-in-headers', () => {
-  const s = new Spectral();
-
-  s.setRules({
-    'valid-oas-example-in-headers': Object.assign(ruleset.rules['valid-oas-example-in-headers'], {
-      recommended: true,
-      type: RuleType[ruleset.rules['valid-oas-example-in-headers'].type],
-    }),
-  });
-
-  test('will pass when simple example is valid', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'string',
-          },
-          example: 'doggie',
-        },
-      },
-    });
-    expect(results).toHaveLength(0);
-  });
-
-  test('will fail when simple example is invalid', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'string',
-          },
-          example: 123,
-        },
-      },
-    });
-    expect(results).toEqual([
-      expect.objectContaining({
-        severity: DiagnosticSeverity.Error,
-        code: 'valid-oas-example-in-headers',
-        message: '"xoxo.example" property type should be string',
-      }),
-    ]);
-  });
-
-  test('will pass when complex example is used ', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'object',
-            properties: {
-              url: {
-                type: 'string',
-              },
-              width: {
-                type: 'integer',
-              },
-              height: {
-                type: 'integer',
-              },
-            },
-            required: ['url'],
-          },
-          example: {
-            url: 'images/38.png',
-            width: 100,
-            height: 100,
-          },
-        },
-      },
-    });
-
-    expect(results).toHaveLength(0);
-  });
-
-  test('will fail when complex example is used ', async () => {
-    const data = {
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'number',
-          },
-          example: 4,
-        },
-        abc: {
-          schema: {
-            type: 'object',
-            properties: {
-              id: {
-                type: 'integer',
-                format: 'int64',
-              },
-              name: {
-                type: 'string',
-              },
-              abc: {
-                type: 'number',
-                example: '5',
-              },
-            },
-            required: ['name'],
-          },
-          example: {
-            name: 'Puma',
-            id: 1,
-          },
-        },
-      },
-    };
-
-    const results = await s.run(data);
-
-    expect(results).toHaveLength(0);
-  });
-
-  test('will error with totally invalid input', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'object',
-            properties: {
-              url: {
-                type: 'string',
-              },
-              width: {
-                type: 'integer',
-              },
-              height: {
-                type: 'integer',
-              },
-            },
-            required: ['url'],
-          },
-          example: {
-            url2: 'images/38.png',
-            width: 'coffee',
-            height: false,
-          },
-        },
-      },
-    });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: 'valid-oas-example-in-headers',
-        message: '"xoxo.example" property should have required property \'url\'',
-        severity: DiagnosticSeverity.Error,
-      }),
-    ]);
-  });
-
-  test('will error with totally invalid input', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'string',
-            format: 'email',
-          },
-          example: 'hello',
-        },
-      },
-    });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: 'valid-oas-example-in-headers',
-        message: '"xoxo.example" property format should match format "email"',
-        path: ['headers', 'xoxo', 'example'],
-        range: expect.any(Object),
-        severity: DiagnosticSeverity.Error,
-      }),
-    ]);
-  });
-
-  test('does not report example mismatches for unknown AJV formats', async () => {
-    const results = await s.run({
-      headers: {
-        xoxo: {
-          schema: {
-            type: 'object',
-            properties: {
-              ip_address: {
-                type: 'integer',
-                format: 'foo',
-              },
-            },
-          },
-          example: 2886989840,
-        },
-      },
-    });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: 'valid-oas-example-in-headers',
-        message: '"xoxo.example" property type should be object',
-        path: ['headers', 'xoxo', 'example'],
-        range: {
-          end: {
-            character: 27,
-            line: 12,
-          },
-          start: {
-            character: 17,
-            line: 12,
-          },
-        },
-        severity: DiagnosticSeverity.Error,
-      }),
-    ]);
-  });
-
-  test.each([['byte', '1'], ['int32', 2 ** 31], ['int64', 2 ** 63], ['float', 2 ** 128]])(
-    'reports invalid usage of %s format',
-    async (format, example) => {
-      const results = await s.run({
-        headers: {
-          xoxo: {
-            schema: {
-              type: ['string', 'number'],
-              format,
-              properties: {
-                ip_address: {
-                  type: 'string',
-                },
-              },
-            },
-            example,
-          },
-        },
-      });
-
-      expect(results).toEqual([
-        expect.objectContaining({
-          severity: DiagnosticSeverity.Error,
-          code: 'valid-oas-example-in-headers',
-          message: `"xoxo.example" property format should match format "${format}"`, // hm, ip_address is likely to be more meaningful no?
-        }),
-      ]);
-    },
-  );
-
-  test.each([['byte', 'MTI3'], ['int32', 2 ** 30], ['int64', 2 ** 40], ['float', 2 ** 64], ['double', 2 ** 1028]])(
-    'does not report valid usage of %s format',
-    async (format, example) => {
-      const results = await s.run({
-        headers: {
-          xoxo: {
-            schema: {
-              type: ['string', 'number'],
-              format,
-              properties: {
-                ip_address: {
-                  type: 'string',
-                },
-              },
-            },
-            example,
-          },
-        },
-      });
-
-      expect(results).toHaveLength(0);
-    },
-  );
-});
+describe(ruleName, () => testRule(ruleName, 'headers'));
