+++ src/rulesets/oas3/__tests__/valid-schema-example-in-content.ts
@@ -1,0 +1,5 @@
+import testRule from './templates/_schema-example';
+
+const ruleName = 'valid-schema-example-in-content';
+
+describe(ruleName, () => testRule(ruleName, 'content'));
